digitizers: Implement new timing trigger logic
Add message input port to digitizers, which is used to connect WR
digitizers (or simulators) and add timing trigger tags to the data
stream.

If the digitizer has a trigger input port (digital or analog)
configured, the detected edges are used together with the timing
messages to generate the tags. If no such port is configured,
the latest timing message is used.

Also removed the time_realignment block, under the assumption that
it is not used anymore in the new architecture, and replaced the old
trigger tag usage in other blocks with the new format.