meson: explicit dependencies to gr libraries
Adds explicit dependencies to gnuradio blocklibs and kernels.
Needed for the correct handling of gnuradio in a nonstandard user
prefix.
This is not very nice imho and we should check with the gnuradio devs
how gnuradio should be imported from meson correctly.
Should probably be a single dependency with user configured modules,
maybe we need to resolve via cmake?