digitizers: Close device when stopping block
start() does the initialization/open and arming, so it seems logical
that stop() is symmetric and does the close(). That should an issue only
if a block is started/stopped often and this causes too much overhead.

Also works around gnuradio/gnuradio#6480 and makes sure every open()
has its close() even if the block is never destroyed.