Picoscope test fixes (#120)
* digitizers: Close device when stopping block

start() does the initialization/open and arming, so it seems logical
that stop() is symmetric and does the close(). That should an issue only
if a block is started/stopped often and this causes too much overhead.

Also works around gnuradio/gnuradio#6480 and makes sure every open()
has its close() even if the block is never destroyed.

* PicoScope: Fix port settings

Add timing port everywhere, fix number of digital ports.

* PicoScope4000a: Arrange tests

Remove unused rapid_block_trigger test, which doesn't apply to PS4000
as there are no digital ports. Make it more clear that there's a
disabled test "streaming_basics" (currently not passing for me).

* Picoscope tests: Fix tag tests

We currently do not connect a timing receiver in the test, so don't
check for trigger tags.

* PicoScope tests: Connect all outputs

Make sure all outputs are connected, otherwise the flowgraph throws
(of outputs are not optional), or hangs (if optional).

* Work around missing error handling when starting digitizer block

When initialization/configuration fails in start(), returning false to
GR is ignored there. GR then calls work() until we claim we're done.
So do that, to abort the flowgraph run when the block is faulty, and
terminate e.g. the unit test.