Picoscope: use offset into driver buffer
This changes the streaming callback and process bulk to also take into
account the data offset in the buffer that the driver is writing the
data into.
Also since the callback is executed on the driver thread, it uses
atomics to exchange the indices.

While this works for now, the picoscope programmer's guide instructs to
copy the data out of the buffer inside of the callback, which suggests
that the data might be overwritten once the callback returns (and
potentially before the processBulk method has actually copied the data).
This should be investigated further.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>