Timing: Align `Event` struct to power of two
Creating a CircularBuffer with trivially copyable Objects of a
non-power-of-two size currently leads to a ring-buffer size that is
incompatible with the way the Circular buffer calculates its internal
indices. See https://github.com/fair-acc/gnuradio4/pull/622 for more
details and proposed fixes.

This should also increase the cache performance for accessing the Event
structure.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>